+++ src/main/java/org/programmers/VoucherManagement/voucher/domain/PercentAmountVoucher.java
@@ -1,25 +1,24 @@
 package org.programmers.VoucherManagement.voucher.domain;
 
+import org.programmers.VoucherManagement.voucher.exception.VoucherException;
+
 import java.util.UUID;
+
+import static org.programmers.VoucherManagement.voucher.exception.VoucherExceptionMessage.NOT_INCLUDE_1_TO_100;
 
 public class PercentAmountVoucher extends Voucher {
     public PercentAmountVoucher(UUID voucherId, DiscountType discountType, DiscountValue discountValue) {
-        super(voucherId, discountType, discountValue);
+        validatePercentDiscountValue(discountValue);
+        this.voucherId = voucherId;
+        this.discountType = discountType;
+        this.discountValue = discountValue;
     }
 
-    @Override
-    public DiscountType getDiscountType() {
-        return discountType;
-    }
 
-    @Override
-    public UUID getVoucherId() {
-        return voucherId;
-    }
-
-    @Override
-    public DiscountValue getDiscountValue() {
-        return discountValue;
+    private void validatePercentDiscountValue(DiscountValue discountValue) {
+        if (discountValue.getValue() < 0 || discountValue.getValue() > 100) {
+            throw new VoucherException(NOT_INCLUDE_1_TO_100);
+        }
     }
 
 }
